requirements.txt
PyYAML>=6.0
click>=8.1
tabulate>=0.9
requests>=2.32
cryptography>=44.0
boto3>=1.35